Give readers the correct enum index, rather than the index from the writer

+++ src/avro/src/types.rs
@@ -82,7 +82,7 @@ pub enum Value {
     /// of its corresponding schema.
     /// This allows schema-less encoding, as well as schema resolution while
     /// reading values.
-    Enum(i32, String),
+    Enum(usize, String),
     /// An `union` Avro value.
     Union(usize, Box<Value>),
     /// An `array` Avro value.
